Added components in about page and more
Card was added showing the head of subsystem, their positions, and a
table consisting of members in a subsystem. These are controlled using
a simple array of js objects. Navigation bar was given the effect where
it hides and docks smoothly when scrolling down and is visible when
scrolling up. Modifications to articles grid (might be ignored in a
future commit)