Major updates
Parallax scroll with cool spring effects in Home page.
A starry background. A whole astra logo svg which is perfect and clear.
Layout of pages with the container so that parallax mouse effects can
be added alongside it. More images uploaded. A lot of CSS changes which
make it fancy.

This was the first commit at which the page was built and deployed onto
gh-pages branch.